requirements.txt
git+https://github.com/huggingface/transformers.git
python-telegram-bot
python-dotenv
